refactoring
rename folder ModelConfiguration to Configurations
